KD: Added new Mappers

+++ src/main/java/Domain/Character.java
@@ -11,7 +11,7 @@ public class Character {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY) //Auto Increment
-	private Integer id;
+	private Long id;
 	
 	
 	private String name;
@@ -22,7 +22,7 @@ public class Character {
 	@ManyToOne
 	private Team team;
 	
-	public Character(Integer id, String name, String weapon, String element, Long level) {
+	public Character(Long id, String name, String weapon, String element, Long level) {
 		super();
 		this.id = id;
 		this.name = name;
@@ -37,10 +37,10 @@ public class Character {
 		this.element = element;
 		this.level = level;
 	}
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
 	public String getName() {
